Get CSS assets to work for middleman local dev server
The original site had several css files. Middleman allows me to
concat these files by using underscore ingrediants in the all.css
file. When reloading the site I see missing images and the content
from the layout.erb in source. The images need to properly be
linked and we need to use the correct erb tags to make this happen.