Add middleman-deploy gem
This commit is to try and use the middelman deploy gem to set up
a workflow for deploying the static assets. The goal is to have
the built assets be in master, while the source branch will be
where I do all my work. These configs have been added to congif.rb
for this reason.