Install foundation to work with middleman for my grid, and styling
I've decided on exposing myself to the foundation framework for this
project. This is because I have experience with bootstrap and I'd
like to gain a new perspective on frontend frameworks. I also like
the idea of using rem instead of pixels. To install this I used the
following blog - http://blachniet.com/2014/04/29/middleman-foundation/