Get bootstrap CSS and JS to work in middleman project
I did this because I need a starting point for using bootstrap in my project.
I still have a bit to learn but I've installed what's needed to use the CSS
and JS.

I learned that Jquery is required for any twitter bootstrap project and the
bootstrap gem does not provide this inclusion.

I got the drop down to work when on mobile screens on the menu bar so I know
that everything is installed correctly.