WIP Style nav bar responsively
I did this because I want to have a style of links for my site that changes
slightly from mobile to larger screens. SASS was hard to get a handle on for
responsive in Bootstrap. I finally found that I could use `@media` to activate
the different styles that are provided by bootstrap.