Finish styling navbar
The styling is finished. I added a brand to the top bar because it feels
better to have one and looks nicer on mobile screens.

I moved the styling into components directory and am storing them there. I
want to be able to quickly find the styling for a given part of my site
without having to think about it too much. I don't know much about the color
yet but I will move on so I don't get stuck styling the nav bar forever.
Hurray!