Add estimated time and style main post list
I did this because I want the post list to contain relevant information. I've
added the ability to truncate the post by adding a `READMORE` flag in the
markdown file.

I've added a JS file for estimating time to read a post. This was an
interesting piece of code and should get a blog post written about it!

I am not done styling but this is a good stopping point. I believe that I
should reevaluate what an MVP blog would be and release it so it's in a live
environment on github pages. Then from there I just continue editing and
updating.