Got default pagination to work
Turns out that if you are using partials you have to declare `pageable: true`
at the top of the main template page ( `index.html.erb` ). Once this happens
all the helper methods become available. You then have to pass them in as
local values if you want to use them. The big kicker for me was having to pass
in page_articles or else it wouldn't use the per_page helper to limit the
pages to my defined 3 articles.