Style individual blog post layout
I did this because I wanted to have a somewhat thin blog layout for posts.
This required special title headings and buttons for switching between
articles. This isn't ideal as I want to display the title of the article but
I am going to save that for after I push it live.