WIP style blog aside and change some file structure issues
I changed the file structure name to partials because I would be confused as
where partials go in the app and would have to look around when trying to
create new ones.

I'm starting to style the blog_aside. There is a lot of information here and I
have to think about how I want to present it. I've gotten tags to look how I
want them and will need to start styling the "random" article list I created.
I spent a lot of time looking into a way to have a page display all tags so I
can limit the number that is shown on the main home page. I might have to
implement this as a custom page. If I do then I might wait until I deploy it
to add that feature.