Style aside content
I want the list of articles to be a random collection of 5 articles. I know
it's pulling form all articles so I put the date with the year so the user can
see how old it is. I am also listing all my tags in a button type format.
Later I think I want to put a link to a list of all articles and maybe limit
the topic tags.