Make pager titles wrap and remove disabled cursor.
I added a width to the pagers on articles because I want long titles to wrap
and not go to far over to the other pagers side. This will make for a better
visual appearance.

I removed the disabled cursor because I would rather not highlight the fact
that it's disabled. The fact that it doesn't turn into a pointer cursor should
help with telling the user this.