Add information in "About" section
I'm trying to define what I want the about section to communicate. I am
including a picture of myself that might change in the future. I want this to
be a place that documents a bit about me as a person and a developer. I don't
want to seem like a know it all or cold so maybe the about section will help.