Identify the current page the user is on
I did this because I want the user to know what page they are currently on
when viewing the website. It will underline the navigation with the current
page. This was hard because it's a static site so the "active" class had to be
added through a helper method instead of Javascript or something done on the
server side. Javascript is used to select the parent of that active link so
the styling of the `li` and the `a` tag could happen.