Add changelog and prepare for 2.0.0 launch
I want to keep a change log of my project for the features that I add and the
bugs that I fix. This way I have something to jog my memory if needed.

I am going to release 2.0.0 of my site and deal with removing the dummy blog
posts after the release. At this point I want to verify that when things are
live on github pages that everthing is working as expected. I believe that I
have some other pages to make but for now those shouldn't stop me from
releasing and starting to write about things.