Fix breaking middleman build with missing scss variables
I did this because the build command was breaking when it was trying to
complie the scss files. This was fixed by included a underscore in the file
names of the files that I didn't want to be included in its own compilation
into scss. The underscored tell Sass not to create a CSS file when it
complies. For whatver reason this still worked on the development environment
but wouldn't work in the build command.