Put correct margin bottom on navbar links
I did this because the `blog` and `about` links were different sizes when next
to each other. The underline takes up 2 px and at the time I Thought it made
sense to use `-2px` thinking it would remove 2 pixels from the margin. Rather,
I was going from a margin bottom of `5px` to `-2px`. This was breaking the
alignment of the links.