fix overlapping footer
I did this because when you looked at the footer links on mobile devices, they
overlapped. The breakpoints for `xs` and `sm` in bootstrap are rather big so
I'm concerned it might look strange on a tablet. For the meantime, until I
make a different design decision for the footer this 3 links then 2 links will
work for me.