Remove favicon directory for build
I did this because I keep getting errors in my build that says it can't find
my favicons. I think this is because the favicon directory isn't being built
so it doesn't know where to look for the images. I'm hoping by putting it in
the images directory I should see the favicons there and the build process
will be able to work correctly.